Duration type changed to double.

+++ lamp_kotlin/src/main/java/digital/lamp/lamp_kotlin/lamp_core/models/DimensionData.kt
@@ -104,7 +104,7 @@ data class LowPowerModeData(
 )
 
 data class TelephonyData(
-    val duration: Int?,
+    val duration: Double?,
     val trace: String?,
     val type: String?
 )
